+++ YART-Lib/include/core/memoryutil.h
@@ -3,10 +3,10 @@
 
 namespace yart
 {
-	void* AllocAligned(size_t size);
+	void* AllocAligned(u64 size);
 
 	template <typename T>
-	T* AllocAligned(size_t count)
+	T* AllocAligned(u64 count)
 	{
 		return (T*)AllocAligned(count * sizeof(T));
 	}
@@ -16,7 +16,7 @@ namespace yart
 	class MemoryArena
 	{
 	public:
-		MemoryArena(size_t blockSize = 262144) // 256 kB
+		MemoryArena(u64 blockSize = 262144) // 256 kB
 			: m_BlockSize(blockSize)
 		{
 		}
@@ -31,14 +31,14 @@ namespace yart
 				FreeAligned(ptr);
 		}
 
-		void* Alloc(size_t nBytes);
+		void* Alloc(u64 nBytes);
 		template <typename T>
-		T* Alloc(size_t count = 1, bool runConstructor = true)
+		T* Alloc(u64 count = 1, bool runConstructor = true)
 		{
 			T* ret = (T*)Alloc(count * sizeof(T));
 			if (runConstructor)
 			{
-				for (size_t i = 0; i < count; i++)
+				for (u64 i = 0; i < count; i++)
 				{
 					new (&ret[i]) T();
 				}
@@ -50,93 +50,93 @@ namespace yart
 		void Reset();
 
 	private:
-		const size_t m_BlockSize;
-		size_t m_BlockOffset = 0, m_CurrentBlockSize = 0;
-		uint8_t* m_CurrentBlock = nullptr;
-		std::vector<std::pair<size_t, uint8_t*>> m_UsedBlocks, m_AvailableBlocks;
+		const u64 m_BlockSize;
+		u64 m_BlockOffset = 0, m_CurrentBlockSize = 0;
+		u8* m_CurrentBlock = nullptr;
+		std::vector<std::pair<u64, u8*>> m_UsedBlocks, m_AvailableBlocks;
 	};
 
 	// Cache friendly generic 2d array
-	template <typename T, int log2BlockSize>
+	template <typename T, i32 log2BlockSize>
 	class BlockedArray
 	{
 	public:
-		BlockedArray(int numX, int numY, const T* arr = nullptr)
+		BlockedArray(i32 numX, i32 numY, const T* arr = nullptr)
 			: m_NumX(numX), m_NumY(numY), m_XBlocks(RoundUp(numX) >> log2BlockSize)
 		{
-			int nAlloc = RoundUp(numX) * RoundUp(numY);
+			i32 nAlloc = RoundUp(numX) * RoundUp(numY);
 			m_Data = AllocAligned<T>(nAlloc);
 
-			for (int i = 0; i < nAlloc; i++)
+			for (i32 i = 0; i < nAlloc; i++)
 			{
 				new (&m_Data[i]) T();
 			}
 
 			if (arr)
 			{
-				for (int y = 0; y < numY; y++)
-					for (int x = 0; x < numX; x++)
+				for (i32 y = 0; y < numY; y++)
+					for (i32 x = 0; x < numX; x++)
 						(*this)(x, y) = arr[y * numX + x];
 			}
 		}
 
 		~BlockedArray()
 		{
-			int nAlloc = RoundUp(m_NumX) * RoundUp(m_NumY);
-			for (int i = 0; i < nAlloc; i++)
+			i32 nAlloc = RoundUp(m_NumX) * RoundUp(m_NumY);
+			for (i32 i = 0; i < nAlloc; i++)
 			{
 				m_Data[i].~T();
 			}
 			FreeAligned(m_Data);
 		}
 
-		constexpr int BlockSize() const
+		constexpr i32 BlockSize() const
 		{
 			return 1 << log2BlockSize;
 		}
 
 		// Rounds x up to the nearest multiple of BlockSize()
-		int RoundUp(int x) const
+		i32 RoundUp(i32 x) const
 		{
 			return (x + BlockSize() - 1) & ~(BlockSize() - 1);
 		}
 
-		int XSize() const
+		i32 XSize() const
 		{
 			return m_NumX;
 		}
 
-		int YSize() const
+		i32 YSize() const
 		{
 			return m_NumY;
 		}
 
-		int Block(int a) const
+		i32 Block(i32 a) const
 		{
 			return a >> log2BlockSize;
 		}
 
-		int Offset(int a) const
+		i32 Offset(i32 a) const
 		{
 			return a & (BlockSize() - 1);
 		}
 
-		T& operator()(int x, int y)
+		T& operator()(i32 x, i32 y)
 		{
-			int bx = Block(x), by = Block(y);
-			int ox = Offset(x), oy = Offset(y);
-			int offset = BlockSize() * BlockSize() * (m_XBlocks * by + bx);
+			i32 bx = Block(x), by = Block(y);
+			i32 ox = Offset(x), oy = Offset(y);
+			i32 offset = BlockSize() * BlockSize() * (m_XBlocks * by + bx);
 			offset += BlockSize() * oy + ox;
 			return m_Data[offset];
 		}
 
-		const T& operator()(int x, int y) const
+		const T& operator()(i32 x, i32 y) const
 		{
 			return const_cast<BlockedArray<T, log2BlockSize>>(*this)(x, y);
 		}
 
 	private:
 		T* m_Data;
-		const int m_NumX, m_NumY, m_XBlocks;
+		const i32 m_NumX, m_NumY, m_XBlocks;
 	};
 }
